3-2 Assignment: User Stories
Overview
After reviewing the Required Resources for this module, use the information in the Supporting Materials section below to complete this activity.

Directions
As the Product Owner for the agile team responsible for creating the SNHU Travel software, you will develop three user stories based on the interactions from the focus group. You must complete the following:

Identify at least five user requests or requirements based on the user interactions in the animation. These should be written by replacing the bracketed text with the necessary information in the following format: As a [type of user], I want to [perform some task] so that I can [achieve some goal].
Next, identify the user requests or requirements based on anticipated effort as large, medium, or small. Prioritize them according to your perceived user needs. Then place these requirements into the Product Backlog worksheet contained in the template provided in the What to Submit section. Order your requirements with the highest priority item first on the list and the lowest priority item at the end.
Ensure your requirements are clear and concise.
Review the User Story Details Example, which is the second worksheet in the template. Then select one large requirement, one medium requirement, and one small requirement from your Product Backlog. Write a detailed user story for each requirement using the User Story Details sheets in the template.
What to Submit
Submit your completed Module Three User Stories Assignment Template for this assignment. Be certain that you have filled in all the different sheets in the template and prioritized the stories in your Product Backlog.