5-2 Assignment: Product Application
Overview
In the course of a product development, the requirements for a project may change and adapt. When this happens, the Product Owner must re-prioritize the Product Backlog so the team can focus on the most important features. Usually, the new features are divided into user stories deemed ready to work by the team. It is important to communicate the new features and adjust priorities to ensure that everyone understands them. In this activity, the Product Owner announces an adjustment to ensure everyone understands the new priorities of the SNHU Travel project’s features. You will assume the role of developer and make the changes to existing code based on the new information from the Product Owner.

If you have questions or issues with this assignment, submit them to the Development Questions Discussion topic. Check the topic periodically and provide answers to your classmates. It is not appropriate to ask for or offer specific code to complete an assignment.

Directions
Review the Product Owner and Scrum-agile Team Animation interaction between the Product Owner and the rest of the Scrum Team. A text version of the animation is available here: Product Owner and Scrum-agile Team Animation Text Version.

As the developer, you have been asked to modify existing code so it will fulfill the new requirements from the Product Owner.

Download the Slide Show ZIP file for the updated Slide Show control the Product Owner gave you in Module Four. Import it into Eclipse.
Update the Slide Show control with pictures and text for the new requirements specified by the Product Owner this week.
Note: Carefully read through the code that you have been given first. Ensure you only change the elements related to the updated requirements from the Product Owner.
Provide specific comments that explain the purpose of lines or sections of your code and detail the approach and method you took to achieve a specific task in the code.
After you have finished completing your assignment, check the code you created. Ask yourself the following questions as you review your work:
Did I revise all appropriate sections of code related to the new requirements?
Is the software functional?
Did I add comments to explain the changes I made and why I made them?
What to Submit
Export your Eclipse Java Project to an executable JAR file. This will include the source code and the image files so the full basic Slide Show control can run. Next, compress (zip) your JAR file and submit your zipped folder to Brightspace.